decouple ClickHouseVersion from Settings
