+++ src/Core/SettingsChangesHistory.cpp
@@ -13,31 +13,7 @@ namespace DB
 
 namespace ErrorCodes
 {
-    extern const int BAD_ARGUMENTS;
     extern const int LOGICAL_ERROR;
-}
-
-ClickHouseVersion::ClickHouseVersion(std::string_view version)
-{
-    Strings split;
-    boost::split(split, version, [](char c){ return c == '.'; });
-    components.reserve(split.size());
-    if (split.empty())
-        throw Exception{ErrorCodes::BAD_ARGUMENTS, "Cannot parse ClickHouse version here: {}", version};
-
-    for (const auto & split_element : split)
-    {
-        size_t component;
-        ReadBufferFromString buf(split_element);
-        if (!tryReadIntText(component, buf) || !buf.eof())
-            throw Exception{ErrorCodes::BAD_ARGUMENTS, "Cannot parse ClickHouse version here: {}", version};
-        components.push_back(component);
-    }
-}
-
-String ClickHouseVersion::toString() const
-{
-    return fmt::format("{}", fmt::join(components, "."));
 }
 
 static void addSettingsChanges(
